+++ packages/dev/core/src/Engines/Native/nativeInterfaces.ts
@@ -6,6 +6,12 @@ import type { Nullable } from "../../types";
 import type { ICanvas, IImage } from "../ICanvas";
 import type { NativeData, NativeDataStream } from "./nativeDataStream";
 
+export type NativeTexture = NativeData;
+export type NativeFramebuffer = NativeData;
+export type NativeVertexArrayObject = NativeData;
+export type NativeProgram = NativeData;
+export type NativeUniform = NativeData;
+
 /** @internal */
 export interface INativeEngine {
     dispose(): void;
@@ -32,15 +38,16 @@ export interface INativeEngine {
     ): void;
     updateDynamicVertexBuffer(vertexBuffer: NativeData, bytes: ArrayBuffer, byteOffset: number, byteLength: number): void;
 
-    createProgram(vertexShader: string, fragmentShader: string): any;
-    getUniforms(shaderProgram: any, uniformsNames: string[]): WebGLUniformLocation[];
-    getAttributes(shaderProgram: any, attributeNames: string[]): number[];
+    createProgram(vertexShader: string, fragmentShader: string): NativeProgram;
+    getUniforms(shaderProgram: NativeProgram, uniformsNames: string[]): WebGLUniformLocation[];
+    getAttributes(shaderProgram: NativeProgram, attributeNames: string[]): number[];
 
-    createTexture(): WebGLTexture;
-    loadTexture(texture: WebGLTexture, data: ArrayBufferView, generateMips: boolean, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
-    loadRawTexture(texture: WebGLTexture, data: ArrayBufferView, width: number, height: number, format: number, generateMips: boolean, invertY: boolean): void;
+    createTexture(): NativeTexture;
+    initializeTexture(texture: NativeTexture, width: number, height: number, hasMips: boolean, format: number, renderTarget: boolean, srgb: boolean): void;
+    loadTexture(texture: NativeTexture, data: ArrayBufferView, generateMips: boolean, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
+    loadRawTexture(texture: NativeTexture, data: ArrayBufferView, width: number, height: number, format: number, generateMips: boolean, invertY: boolean): void;
     loadRawTexture2DArray(
-        texture: WebGLTexture,
+        texture: NativeTexture,
         data: Nullable<ArrayBufferView>,
         width: number,
         height: number,
@@ -49,14 +56,14 @@ export interface INativeEngine {
         generateMipMaps: boolean,
         invertY: boolean
     ): void;
-    loadCubeTexture(texture: WebGLTexture, data: Array<ArrayBufferView>, generateMips: boolean, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
-    loadCubeTextureWithMips(texture: WebGLTexture, data: Array<Array<ArrayBufferView>>, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
-    getTextureWidth(texture: WebGLTexture): number;
-    getTextureHeight(texture: WebGLTexture): number;
-    copyTexture(desination: Nullable<WebGLTexture>, source: Nullable<WebGLTexture>): void;
-    deleteTexture(texture: Nullable<WebGLTexture>): void;
+    loadCubeTexture(texture: NativeTexture, data: Array<ArrayBufferView>, generateMips: boolean, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
+    loadCubeTextureWithMips(texture: NativeTexture, data: Array<Array<ArrayBufferView>>, invertY: boolean, srgb: boolean, onSuccess: () => void, onError: () => void): void;
+    getTextureWidth(texture: NativeTexture): number;
+    getTextureHeight(texture: NativeTexture): number;
+    copyTexture(desination: NativeTexture, source: NativeTexture): void;
+    deleteTexture(texture: NativeTexture): void;
     readTexture(
-        texture: WebGLTexture,
+        texture: NativeTexture,
         mipLevel: number,
         x: number,
         y: number,
@@ -71,14 +78,12 @@ export interface INativeEngine {
     resizeImageBitmap(image: ImageBitmap, bufferWidth: number, bufferHeight: number): Uint8Array;
 
     createFrameBuffer(
-        texture: WebGLTexture,
+        texture: Nullable<NativeTexture>,
         width: number,
         height: number,
-        format: number,
         generateStencilBuffer: boolean,
-        generateDepthBuffer: boolean,
-        generateMips: boolean
-    ): WebGLFramebuffer;
+        generateDepthBuffer: boolean
+    ): NativeFramebuffer;
 
     getRenderWidth(): number;
     getRenderHeight(): number;
